Add HDMI support
This change checks whether the HDMI cable is plugged or not
and starts RetroArch with the corresponding configuration
(display resolution and audio device must be set via
retroarch.cfg). The retroarch config file is selected
based on the current display resolution, and supports
HDMI (720p) and internal displays for powkiddy v2 and v3.

Weston is configured to support the internal display and
HDMI at 720p@60.

Audio path is also properly set. It's worth noting that HDMI
audio output is 'hw:0,1', so RetroArch should be configured
accordingly.