Add sync mount option and call poweroff on exit RA
Read/Write filesystems are now mounted with `sync` option
to make sure no changes are lost when powering down the
system.

In addition to this, `poweroff` command is called when exiting
RetroArch, so there is a proper way of shutting down the
system, as pushing the button just cuts the power.