Add LICENSE file to the root directory
This License file is in essence the same license file provided
by Buildroot project. As Ruka CFW is based on Buildroot, it
makes sense to keep the same license just changing the name of
the project to cover all the files under Ruka CFW and not only
those under Buildroot.